:construction: Add torpedo_accuracy_mod

+++ crates/fleethub-core/src/master_data/master_ship.rs
@@ -39,7 +39,7 @@ pub struct MasterShip {
     pub los: StatInterval,
     pub luck: StatInterval,
     #[serde(default)]
-    pub torpedo_accuracy: u8,
+    pub torpedo_accuracy: u16,
     pub speed: u8,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub range: Option<u8>,
